refkit-ostree: swallow most code from the initramfs ostree module.
Added refkit-ostree, a consolidated script where we try to collect
in a single place all the extra tricks we need to pull to get ostree-
based image updates working on top of refkit.

Moved practically all the code from the original initramfs ostree
module to refkit-ostree. The initramfs module now picks refkit-ostree
from the first deployment it finds and then executes it to do all the
necessary things to prepare an ostree-based refkit image to boot.

refkit-ostree script has a bunch of modes of operation, including:

  - initramfs-prepare-root: do the deed for the ostree module
  - patch-proc-cmdline: patch /proc/cmdline with the ostree argument
  - mount-efi: mount the EFI partition on /boot/efi/EFI
  - fetch-updates: fetch pending updates (without applying them)
  - apply-updates: apply any pending updates (without fetching)
  - update-uefi-app: update the UEFI app if necessary
  - upgrade: fetch + apply + apply UEFI app updates
  - rollback: not implemented
  - running-version: show the currently running deployment
  - latest-version: show the latest available deployment

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>