refkit-ostree-update: added C refkit-ostree, split out update.
Implemented most of the functionality needed by the initramfs refkit
ostree module. Split up the functionality so that bits depending on
libostree are now only available in the original refkit-ostree-update
binary. The rest of the reduced subset is available in a new binary
as ${bindir}/refkit-ostree.

Invoking refkit-ostree with the --prepare-root option from an initramfs
can be used to discover and prepare the latest ostree deployment root
for booting. It essentially performs the same operations as the original
script or the refkit-ostree script when run with the initramfs-prepare
argument.

The original refkit-ostree script has been moved out of the way, under
${datadir}/refkit-ostree/scripts.

Signed-off-by: Krisztian Litkey <kli@iki.fi>