scripts, services, hooks: added post-apply reboot hook and service.
Added a post-apply hook to request a reboot after a successful
system update. The reboot will activate and boot into the latest
ostree deployment.

Entities that need to prevent (= delay) the shutdown to a more
convenient time in the future should use systemd-inhibit or the
corresponding systemd interfaces for doing so. For instance, if
you have an interactive shell while the updater is running and
you don't want the system to be rebooted under you in case an
update gets pulled in, you should do a

    systemd-inhibit --what=shutdown $SHELL

and exit that new shell, once you're finished and want to allow
the reboot/update to proceed.

Because of known limitations in systemd-inhibit (see below) also
added a systemd service which waits (currently indefinitely) for
all systemd-inhibit shutdown locks then reboots the system.

This is necessary as systemd-inhibit shutdown locks are ignored
both for privileged users (explictly ignore for root) and also
for non-interactive invocations of systemctl (to provide backward-
compatible behavior for scripts that explicitly run shutdown or
reboot which are now just symlinked to systemctl).

These are all known limitations of systemd-logind which hosts the
inhibitor lock code and API. There are feature requests open to add
a logind.conf configuration entry to override the default behavior
and obey locks for root and scripts as well.

See, for instance, https://github.com/systemd/systemd/issues/949,
or https://github.com/systemd/systemd/issues/2680 for more details.

The actual inhibitable reboot is implemented by the newly added
straightforward inhibitable-reboot script.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>